DB methods, Routing and basic version of a Single Post Page
Added DB methods for single post, replies and posts. Modified custom object to be able to sort by timestamp, and keep track of the ID of the post. Created a basic version of a single post view, and dealt with routing when a post in the thread view is clicked.